wip: linting and correcting error classes
