+++ errors/UnauthorizedError.js
@@ -1,0 +1,10 @@
+class UnauthorizedError extends Error {
+  constructor(message) {
+    super(message);
+    this.statusCode = 401;
+  }
+}
+
+module.exports = {
+  UnauthorizedError,
+};
